Refactor structure of shape module
