+++ packages/punchcard/lib/dynamodb/expression/compile-context.ts
@@ -1,5 +1,5 @@
 import { Writer } from "../../shape/mapper/mapper";
-import { Type } from "../../shape/types/type";
+import { Type } from "../../shape/type";
 
 export interface CompileContext {
   readonly writer: Writer<AWS.DynamoDB.AttributeValue>;
