+++ packages/punchcard/lib/shape/hash.ts
