+++ packages/punchcard/lib/shape/json/index.ts
@@ -1,2 +1,3 @@
 export * from './mapper';
 export * from './path';
+export * from './util';
