+++ packages/punchcard/lib/shape/kind.ts
