+++ packages/punchcard/lib/shape/primitive.ts
@@ -1,5 +1,5 @@
-import { DynamoPath } from '../../dynamodb/expression/path';
-import { JsonPath } from '../json/path';
+import { DynamoPath } from '../dynamodb/expression/path';
+import { JsonPath } from './json/path';
 import { Kind } from './kind';
 import { Type } from './type';
