--- packages/punchcard/lib/shape/types/index.ts
@@ -1,16 +1,0 @@
-export * from './array';
-export * from './binary';
-export * from './boolean';
-export * from './dynamic';
-export * from './hash';
-export * from './kind';
-export * from './map';
-export * from './number';
-export * from './optional';
-export * from './primitive';
-export * from './set';
-export * from './string';
-export * from './struct';
-export * from './timestamp';
-export * from './type';
-export * from './typed-set';
